requirements.txt
opencv-python
pytesseract
fastapi
uvicorn
python-multipart
pillow
streamlit
requests
transformers
torch